filtro

+++ esercitazioni/86-RazorApp/models/Prodotti.cs
@@ -2,7 +2,7 @@ public class Prodotto
 {
     public int Id {get;set;}
     public string Nome { get; set; }
-    public int Prezzo { get; set; }
+    public decimal Prezzo { get; set; }
     public string Dettaglio { get; set; }
     public string Immagine { get; set; }
 }
